Move modding to its own controller
